bump core submodule ref to allow adding extra gce pd, with full control
This again uses the raw `disks_struct` param.

We could have used the `disks` param to launch a new gce instance with
extra pd, but the pd will have auto-generated device name and not be
auto-deleted.

We could have modified the `disks` param to allow setting device name
and toggling auto-delete flag, but the logic of existing code seems a
little complicated. To revisit.